# swagger.yaml
openapi: 3.0.0
info:
    title: 'API Documentation'
    version: 1.0.0
    description: 'API documentation for plans, workouts, and exercises.'
servers:
    -
        url: 'http://localhost:3010'
paths:
...
    /register:
        post:
            summary: 'Register a new user'
            description: 'This endpoint allows a user to register by providing a username and password.'
            tags: [Authentication]
            requestBody: {required: true, content: {application/json: {schema: {type: object, properties: {username: {type: string}, password: {type: string}}}}}}
            responses: {'201': {description: 'User successfully registered', content: {application/json: {schema: {type: object, properties: {message: {type: string, example: 'User registered successfully.'}, token: {type: string, description: 'The JWT token for the authenticated user.'}}}}}}, '400': {description: 'Bad request (e.g., missing username or password)'}, '500': {description: 'Internal server error'}}
    /login:
        post:
            summary: 'Login a user'
            description: 'This endpoint allows a user to login using their username and password.'
            tags: [Authentication]
            requestBody: {required: true, content: {application/json: {schema: {type: object, properties: {username: {type: string}, password: {type: string}}}}}}
            responses: {'200': {description: 'Login successful', content: {application/json: {schema: {type: object, properties: {message: {type: string, example: 'Login successful.'}, token: {type: string, description: 'The JWT token for the authenticated user.'}}}}}}, '400': {description: 'Invalid username or password'}, '500': {description: 'Internal server error'}}
    /renew:
        post:
            summary: 'Renew the access token'
            description: 'This endpoint allows a user to renew their access token using a valid refresh token.'
            tags: [Authentication]
            requestBody: {required: true, content: {application/json: {schema: {type: object, properties: {token: {type: string, description: 'The refresh token to be used for generating a new access token.'}}}}}}
            responses: {'200': {description: 'Token successfully renewed', content: {application/json: {schema: {type: object, properties: {message: {type: string, example: 'Token renewed successfully.'}, token: {type: string, description: 'The newly generated JWT token.'}}}}}}, '400': {description: 'Missing or invalid token'}, '401': {description: 'Invalid or expired token'}}
...